+++ MLEM.Extended/Extensions/TextureExtensions.cs
@@ -1,6 +1,8 @@
 using MLEM.Extensions;
 using MLEM.Textures;
-using MonoGame.Extended.TextureAtlases;
+using MonoGame.Extended.Graphics;
+using NinePatch = MLEM.Textures.NinePatch;
+using ExtNinePatch = MonoGame.Extended.Graphics.NinePatch;
 
 namespace MLEM.Extended.Extensions {
     /// <summary>
@@ -9,38 +11,29 @@ namespace MLEM.Extended.Extensions {
     public static class TextureExtensions {
 
         /// <summary>
-        /// Converts a MLEM <see cref="NinePatch"/> to a MonoGame.Extended <see cref="NinePatchRegion2D"/>.
+        /// Converts a MLEM <see cref="Textures.NinePatch"/> to a MonoGame.Extended <see cref="ExtNinePatch"/>.
         /// </summary>
         /// <param name="patch">The nine patch to convert</param>
         /// <returns>The converted nine patch</returns>
-        public static NinePatchRegion2D ToExtended(this NinePatch patch) {
-            return new NinePatchRegion2D(patch.Region.ToExtended(), patch.Padding.Left.Floor(), patch.Padding.Top.Floor(), patch.Padding.Right.Floor(), patch.Padding.Bottom.Floor());
+        public static ExtNinePatch ToExtended(this NinePatch patch) {
+            return patch.Region.ToExtended().CreateNinePatch(patch.Padding.Left.Floor(), patch.Padding.Top.Floor(), patch.Padding.Right.Floor(), patch.Padding.Bottom.Floor());
         }
 
         /// <summary>
-        /// Converts a MLEM <see cref="TextureRegion"/> to a MonoGame.Extended <see cref="TextureRegion2D"/>.
+        /// Converts a MLEM <see cref="TextureRegion"/> to a MonoGame.Extended <see cref="Texture2DRegion"/>.
         /// </summary>
         /// <param name="region">The nine patch to convert</param>
         /// <returns>The converted nine patch</returns>
-        public static TextureRegion2D ToExtended(this TextureRegion region) {
-            return new TextureRegion2D(region.Name, region.Texture, region.U, region.V, region.Width, region.Height);
+        public static Texture2DRegion ToExtended(this TextureRegion region) {
+            return new Texture2DRegion(region.Texture, region.U, region.V, region.Width, region.Height, region.Name);
         }
 
         /// <summary>
-        /// Converts a MonoGame.Extended <see cref="NinePatchRegion2D"/> to a MLEM <see cref="NinePatch"/>.
-        /// </summary>
-        /// <param name="patch">The nine patch to convert</param>
-        /// <returns>The converted nine patch</returns>
-        public static NinePatch ToMlem(this NinePatchRegion2D patch) {
-            return new NinePatch(((TextureRegion2D) patch).ToMlem(), patch.LeftPadding, patch.RightPadding, patch.TopPadding, patch.BottomPadding);
-        }
-
-        /// <summary>
-        /// Converts a MonoGame.Extended <see cref="TextureRegion2D"/> to a MLEM <see cref="TextureRegion"/>.
+        /// Converts a MonoGame.Extended <see cref="Texture2DRegion"/> to a MLEM <see cref="TextureRegion"/>.
         /// </summary>
         /// <param name="region">The nine patch to convert</param>
         /// <returns>The converted nine patch</returns>
-        public static TextureRegion ToMlem(this TextureRegion2D region) {
+        public static TextureRegion ToMlem(this Texture2DRegion region) {
             return new TextureRegion(region.Texture, region.Bounds) {Name = region.Name};
         }